websocket-service/requirements.txt
Flask==1.1.2
Flask_SocketIO==4.3.2
requests==2.24.0
kafka-python==2.0.1
loguru==0.5.3